Taking Flight to the Land of Off-Heap, Act I: Page Meta-Data

+++ community/io/src/main/java/org/neo4j/io/pagecache/PageCache.java
@@ -107,7 +107,7 @@ public interface PageCache extends AutoCloseable
     /**
      * The max number of cached pages.
      **/
-    int maxCachedPages();
+    long maxCachedPages();
 
     /**
      * Get the {@link FileSystemAbstraction} that represents the filesystem where the paged files reside.
